libserver: fix type passed into ec_log
ECSession.cpp: In member function "ValidateSSOData_NTLM":
ECLogger.h:335:66: warning: field precision specifier ".*" expects
argument of type "int", but argument 3 has type "ssize_t {aka long int}"

Fixes: bp/8.3.1-42-ga1c5adb0

+++ provider/libserver/ECSession.cpp
@@ -1383,7 +1383,7 @@ retry:
 		 * (unlikely), or ntlm_auth found some reason not to complete,
 		 * like /var/lib/samba/winbindd_privileged being inaccessible.
 		 */
-		ec_log_err("ntlm_auth returned generic error \"%.*s\"", bytes, buffer);
+		ec_log_err("ntlm_auth returned generic error \"%.*s\"", static_cast<int>(bytes), buffer);
 		return er;
 	} else if (buffer[0] == 'T' && buffer[1] == 'T') {
 		// Try This
